Check playlist order on subscribe
